requirements.txt
anyio==3.6.2
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
cloudinary==1.32.0
dnspython==2.3.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
mongoengine==0.26.0
pycodestyle==2.10.0
pydantic==1.10.5
pymongo==4.3.3
python-dotenv==0.21.1
python-multipart==0.0.5
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.14
uvicorn==0.20.0
